Checkpoint

+++ src/components/ui/forms/EditProfile.jsx
@@ -1,14 +1,13 @@
 import { useState, useContext } from "react";
-import { useLoaderData, useNavigate } from "react-router-dom"
+import { useNavigate } from "react-router-dom";
 import { ActiveUserContext } from "../../ActiveUserContext";
-import "../../Style/ButtonStyle/button.css"
-import "../../Style/EditProfileStyle/EditProfile.css"
+import "../../Style/ButtonStyle/button.css";
+import "../../Style/EditProfileStyle/EditProfile.css";
 
 export const EditProfile = () => {
-  const { users } = useLoaderData()
   const navigateTo = useNavigate();
   const [activeUser, setActiveUser] = useContext(ActiveUserContext);
-  const currentProfile = activeUser
+  const currentProfile = activeUser;
   const id = currentProfile[0].id;
   const currentImage = currentProfile[0].image;
   const userName = currentProfile[0].name;
@@ -22,20 +21,22 @@ export const EditProfile = () => {
 
   const handleEditMode = () => {
     setEditMode((current) => !current);
-  }
+  };
   const back = () => {
     setEditMode((current) => !current);
-  }
+  };
 
   const handleDelete = () => {
-    fetch(`http://localhost:3000/users/${id}`, {
-      method: "DELETE",
-    }).then(() => {
-      setActiveUser([{id: 0}])
-      window.alert("User deleted");
-      navigateTo("/")
-    });
-  }
+    const result = confirm("Are you sure?");
+    if (result == true) {
+      fetch(`http://localhost:3000/users/${id}`, {
+        method: "DELETE",
+      }).then(() => {
+        setActiveUser([{ id: 0 }]);
+        navigateTo("/");
+      });
+    }
+  };
 
   const handleSubmit = (e) => {
     e.preventDefault();
@@ -51,15 +52,21 @@ export const EditProfile = () => {
       body: JSON.stringify(user),
     }).then(() => {
       window.alert("User edited");
-      sessionStorage.setItem('activeUser', JSON.stringify({
-        "id": id
-      }))
-      setEditMode((current) => !current);
-    })
+      sessionStorage.setItem(
+        "activeUser",
+        JSON.stringify({
+          id: id,
+          name: name,
+          password: password,
+          image: image,
+        })
+      );
+      navigateTo(0);
+    });
   };
 
   if (editMode == true) {
-    return(
+    return (
       <>
         <div className="profile">
           <div className="card">
@@ -70,14 +77,18 @@ export const EditProfile = () => {
                   placeholder="Username"
                   required
                   value={name}
-                  onChange={(newUsername) => setNewUserName(newUsername.target.value)}
+                  onChange={(newUsername) =>
+                    setNewUserName(newUsername.target.value)
+                  }
                 ></input>
                 <input
                   className="textInput"
                   placeholder="Password"
                   required
                   value={password}
-                  onChange={(newPassword) => setNewPassword(newPassword.target.value)}
+                  onChange={(newPassword) =>
+                    setNewPassword(newPassword.target.value)
+                  }
                 ></input>
                 <input
                   className="textInput"
@@ -88,7 +99,9 @@ export const EditProfile = () => {
                   }
                 ></input>
                 <div className="buttonContainer">
-                  <button className="button" onClick={back}>Back</button>
+                  <button className="button" onClick={back}>
+                    Back
+                  </button>
                   <button className="button" type="submit">
                     Submit
                   </button>
@@ -98,28 +111,27 @@ export const EditProfile = () => {
           </div>
         </div>
       </>
-    )
+    );
   } else {
-    return(
+    return (
       <>
         <div className="profile">
           <div className="card">
-          <img src={currentImage} ></img>
-          <div className="infoBox">
-            <h1>{userName}</h1>
-            <div className="buttonContainer">
-                    <button className="button" onClick={handleDelete}>
-                      Delete
-                    </button>
-                    <button className="button" onClick={handleEditMode}>
-                      Edit
-                    </button>
-                  </div>
-          </div>
+            <img src={currentImage}></img>
+            <div className="infoBox">
+              <h1>{userName}</h1>
+              <div className="buttonContainer">
+                <button className="button" onClick={handleDelete}>
+                  Delete
+                </button>
+                <button className="button" onClick={handleEditMode}>
+                  Edit
+                </button>
+              </div>
+            </div>
           </div>
         </div>
       </>
-    )
+    );
   }
-
 };
